v.0.4.0 Tag editing Style editing Starting on intro page

+++ src/renderer/components/Tags.vue
@@ -3,16 +3,20 @@
         <table>
             <tr>
                 <th>Name</th>
-                <th>Style</th>
+                <th style="width:40%">Style</th>
+                <th style="width:30%">Preview</th>
                 <th></th>
             </tr>
-            <tr v-for="(tag, i) in tags">
-                <td><input type="text" v-model="tag.name"></td>
-                <td><textarea v-model="tag.style"></textarea></td>
+            <tr v-for="(tag, i) in tags" @keyup="tag.saved = false">
                 <td>
-                    <button @click="saveTag(tag)">Save</button>
-                    <button @click="deleteTag(tag.tag_id)">Delete</button>
-                    <div>{{ tag.message }}</div>
+                    <input style="width: 120px" type="text" v-model="tag.name"><br>
+                </td>
+                <td><textarea v-model="tag.style"></textarea></td>
+                <td>This is an <span :title="tag.name" :style="tag.style">example of how your final style will look</span>
+                    when it appears in your journal.</td>
+                <td>
+                    <button v-if="!tag.saved" @click="saveTag(tag)">Save</button>
+                    <button @click="deleteTag(tag.tag_id, $event)">Delete</button>
                 </td>
             </tr>
             <button @click="newTag">Add New</button>
@@ -43,7 +47,11 @@
 
     textarea {
         height: 100px;
-        width: 400px;
+        width: 100%;
+    }
+
+    input {
+        font-size: 1em;
     }
 </style>
 
@@ -59,8 +67,13 @@
       this.$db.all('SELECT * FROM tags ORDER BY name ASC')
         .then((rows) => {
           rows.forEach((row) => {
-            row.message = ''
-            this.tags.push(row)
+            this.tags.push({
+              tag_id: row.tag_id,
+              name: row.name,
+              type: 'inline',
+              style: row.style,
+              saved: true
+            })
           })
         })
     },
@@ -73,7 +86,7 @@
               name: '',
               type: 'inline',
               style: '',
-              message: ''
+              saved: true
             })
           })
           .catch((error) => {
@@ -83,24 +96,29 @@
       saveTag (tag) {
         this.$db.run('UPDATE tags SET name = ?, type = ?, style = ? WHERE tag_id = ?', [tag.name, tag.type, tag.style, tag.tag_id])
           .then(() => {
-            tag.message = 'Tag saved'
-            setTimeout(() => {
-              tag.message = ''
-            }, 1500)
+            tag.saved = true
           })
       },
-      deleteTag (id) {
-        this.$db.run('DELETE FROM tags WHERE tag_id = ?', [id])
-          .then((ret) => {
-            if (ret.changes) {
-              this.tags = this.tags.filter(tag => {
-                return tag.tag_id !== id
-              })
-            }
-          })
-          .catch((error) => {
-            console.log(error)
-          })
+      deleteTag (id, event) {
+        let confirm = 'Are you sure?'
+        if (event.target.innerText === confirm) {
+          this.$db.run('DELETE FROM tags WHERE tag_id = ?', [id])
+            .then((ret) => {
+              if (ret.changes) {
+                this.tags = this.tags.filter(tag => {
+                  return tag.tag_id !== id
+                })
+              }
+            })
+            .catch((error) => {
+              console.log(error)
+            })
+        } else {
+          event.target.innerText = confirm
+          setTimeout(() => {
+            event.target.innerText = 'Delete'
+          }, 3000)
+        }
       },
       exit () {
         this.$router.go(-1)
